==> 3-processing.ipynb <==
# Processing the data

# standard library imports
import csv
import datetime as dt
import json
import os
import statistics
import time

# third-party imports
import numpy as np
import pandas as pd
pd.set_option("max_columns", 100)

## 1) Outliers
So far I have cleaned the empty fields, but there are still some fields with wrong values or that presents too strong biases.

steam_clean = pd.read_csv('data/steam_clean.csv')

def number_uniques(table):
    for col in table.columns:
        n = len(pd.unique(table[col]))
        
        print('Column', col, 'has', n, 'different values.')

number_uniques(steam_clean)

steam_clean.dtypes

With that overview of the unique values, I need to understand them:

def bias_review(table):
    cols_to_review = []
    size = table.shape[0]
    frontier = size*0.9
    
    for col in table.columns:
        mcv = table[col].value_counts(dropna=False).max() #most common value
        if (mcv > frontier):
            cols_to_review.append(col)
        print(col, mcv)
    return cols_to_review

steam_clean['appid'].value_counts().max()

mcv = steam_clean['genre'].value_counts()[0]
size = steam_clean.shape[0]
frontier = size*0.9
(mcv > frontier)
frontier

cols_to_review = bias_review(steam_clean)
cols_to_review

The function above gives me a list of columns to check

**Types**

steam_clean.type.value_counts()

There is only one field with hardware. That means that line can be dropped.

steam_clean = steam_clean.loc[steam_clean["type"] != 'hardware']

Now that all lines are confirmed as games, I don't need this column.

cols_to_drop = ['type']

**demos**

steam_clean.demos.value_counts()

Even though this column appear as biased, the data it can offer is useful enough to keep it.

**userscore**

steam_clean.userscore.value_counts()

The amount of games without userscore makes this column usless, I will need to drop this and rely on other columns with similar meaning.

cols_to_drop.append('userscore')

**average_2weeks**

steam_clean.average_2weeks.value_counts()

I think this column and the other average and median columns show the amount of time played. As I cannot find the actual explanation or the units for this column, and seeing how biased are all 4 columns, I won't use them.

cols_to_drop.append('average_forever')
cols_to_drop.append('average_2weeks')
cols_to_drop.append('median_forever')
cols_to_drop.append('median_2weeks')

**discount**

steam_clean.discount.value_counts()

This column shows the actual discount for the games. Out of sales seasons (like when this sample was taken) the discounts aren't that common. In this case, I think the column can be dropped.

cols_to_drop.append('discount')

cols_to_drop

steam_clean = steam_clean.drop(cols_to_drop, 1)

Besides the biased columns, I need to look at a few more that I know show problems from the initial exploration:

**Age**

steam_clean.required_age.value_counts()

This field presents a lot of issues from the get go, the main ones being the two string fields: `16+` and `javascript`.  
Once I fix them, I can turn everithing into integers.
First I drop the javascript line as it has no numerical correspondence.

steam_clean = steam_clean.loc[steam_clean["required_age"] != "javascript:ToggleCheckbox('checkbox_app_game_ratings_pegi_use_age_gate_');"]

steam_clean["required_age"].replace({"16+": 16}, inplace=True)

steam_clean["required_age"] = steam_clean["required_age"].astype('float')

steam_clean.required_age.value_counts()

After the cleanup, I can see that the amount of 0 values might be too high, so I flag this column for deletion.

cols_to_drop = ['required_age']

**developers** and **publishers**  
These two columns show the same format:

steam_clean.developers.value_counts()

steam_clean.publishers.value_counts()

When the value is an empty string in the publisher column, it means that the developers didn't have a publisher behind them. I will leave the treatment of these two columns for when I need to encode them.

## 2) Dictionaries
There are some columns with fields made of dictionaries and while that may save space, it makes reading the information much more difficult. To handle the information inside those columns, I need to create anew columnes based on the contents of their fields:

import ast

def treat_dict(column):
    df = column.map(ast.literal_eval)
    df = df.apply(pd.Series)
    return df

platforms = treat_dict(steam_clean.platforms)
cols_to_drop.append('platforms')

platforms

platforms.windows.value_counts()

platforms.mac.value_counts()

platforms.linux.value_counts()

releases = treat_dict(steam_clean.release_date)
releases.rename(columns={'date': 'release'}, inplace=True)
cols_to_drop.append('release_date')
releases

recommendations = treat_dict(steam_clean.recommendations)
cols_to_drop.append('recommendations')
recommendations.rename(columns={'total': 'amount_recs'}, inplace=True)
recommendations

tags = treat_dict(steam_clean.tags)
tags

This column creates too many different columns as it is made from the list of tags users have associated to the games. I think I won't need the column at all.

cols_to_drop.append('tags')

**genre**  
This columns show the genres associated to a game in the form of a comma separated list of variable amount of genres.
I want to keep only the most important ones.

steam_clean.genre.value_counts()

genres = steam_clean.genre.str.split(', ')
genres = pd.DataFrame(genres.tolist())
genres = genres[[0, 1, 2]]
genres.rename(columns={0: 'genre1', 1:'genre2', 2:'genre3'}, inplace=True)
cols_to_drop.append('genre')
genres

Those `None` fields are NaNs apparently, so that has to be fixed.  
My way of fixing it will be to give them the actual `None` value as a string.

genres.isna().sum()

genres.genre3.value_counts(dropna=False)

genres = genres.fillna('None')

genres.genre3.value_counts(dropna=False)

## 3) Reforming the dataframe
After all the processing has been done, I need to get all the data again in a dataframe, but first I need to remove the columns that will be replaced.

steam_dropped = steam_clean[cols_to_drop]
steam_clean = steam_clean.drop(cols_to_drop, 1)
steam_clean.shape

steam_processed = pd.concat([steam_clean, platforms, releases.release, genres], axis=1)

steam_processed

steam_processed.to_csv('data/steam_processed.csv', index=False)

**DATABASE**

import pymysql
from sqlalchemy import create_engine
import getpass  # To get the password without showing the inputb

password = getpass.getpass()

connection_string = 'mysql+pymysql://root:'+password+'@localhost/steam'
engine = create_engine(connection_string)

steam_processed.to_sql('steam_processed', con=engine)
