==> 5-hypothesis.ipynb <==
# Hypothesis Testing

# standard library imports
import csv
import datetime as dt
import json
import os
import statistics
import time

# third-party imports
import numpy as np
import pandas as pd
pd.set_option("max_columns", 100)
import scipy.stats
from scipy.stats import ttest_1samp


steam_data = pd.read_csv('data/steam_processed.csv')

steam_data

~~I have some understanding of the industry, and I want to test if more than 50% of the RPG games that are free have DLCs.~~  
I have some understanding of the industry, and I want to test if RPG games that are free have more than 500 positive reviews.  

The null hypothesis in this case would be what was said above.  
The alternative hypothesis is that not being true.
My significance level is the usual 0.05, which means I am 95% confident on my hypothesis for any sample.

## 1) Selecting data


free_rpgs = steam_data[((steam_data['genre1']=='RPG')|(steam_data['genre2']=='RPG')|(steam_data['genre3']=='RPG'))&steam_data['is_free']]['positive']
free_rpgs

## 2) Testing the hypothesis

stat, pval = ttest_1samp(free_rpgs, 500)
print('stat is  ', stat)
print('pvalue for the two-tailed test is ', pval)

In this case, the two tailed test indicates that I need to reject my null hypothesis.
